+++ scripts/unicode.py
@@ -188,11 +188,10 @@ def emit_enums(f, script_list, extension_list, longforms, intersections):
     Emit the Script and ScriptExtension enums as well as any related utility functions
     """
     f.write("""
-use core::convert::TryFrom;
 #[derive(Clone, Copy, PartialEq, Eq, Debug, Hash)]
 #[non_exhaustive]
 #[allow(non_camel_case_types)]
-/// A value of the Script property
+/// A value of the `Script` property
 pub enum Script {
     /// Unknown script
     Unknown,
@@ -202,11 +201,11 @@ pub enum Script {
     f.write("""}
 #[derive(Clone, Copy, PartialEq, Eq, Debug, Hash)]
 #[non_exhaustive]
-/// A value for the Script_Extension property
+/// A value for the `Script_Extension` property
 ///
-/// Script_Extension is one or more Script
+/// [`ScriptExtension`] is one or more [`Script`]
 ///
-/// This is essentially an optimized version of Vec<Script>,
+/// This is essentially an optimized version of `Vec<Script>`,
 /// optimized by script sets and intersections actually present in Unicode.
 pub enum ScriptExtension {
     /// A single script
@@ -217,23 +216,8 @@ pub enum ScriptExtension {
         f.write("    /// %s\n    %s,\n" % (longform, "".join(ext)))
     f.write("""}
 
-impl From<Script> for ScriptExtension {
-    fn from(script: Script) -> Self {
-        ScriptExtension::Single(script)
-    }
-}
-
-impl TryFrom<ScriptExtension> for Script {
-    type Error = ();
-    fn try_from(ext: ScriptExtension) -> Result<Self, ()> {
-        match ext {
-            ScriptExtension::Single(s) => Ok(s),
-            _ => Err(())
-        }
-    }
-}
-
 impl Script {
+    #[inline]
     pub(crate) fn inner_full_name(self) -> &'static str {
         match self {
             Script::Unknown => "Unknown",
@@ -243,6 +227,7 @@ impl Script {
     f.write("""        }
     }
 
+    #[inline]
     pub(crate) fn inner_short_name(self) -> &'static str {
         match self {
             Script::Unknown => "",
